DEV: Snaketool CLI utility functions now a separate module.
