+++ setup.py
@@ -55,6 +55,7 @@ setup(
     data_files=data_files,
     package_data=package_data,
     install_requires=[
+        "snaketool-utils>=0.0.2",
         "snakemake>=7.14.0,<=7.26.0",
         "pyyaml>=6.0",
         "Click>=8.1.3",
